+++ packages/editor/editor-common/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-common
 
+## 32.4.2
+- [patch] [4eb1af2892](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4eb1af2892):
+
+  - ED-6265 fix external image call to media for dimensions
+
 ## 32.4.1
 - [patch] [42b78a6133](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/42b78a6133):