+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 3.5.0
+- [minor] [9bb0ecb48a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9bb0ecb48a):
+
+  - Support wiki to smart link
+
 ## 3.4.0
 - [minor] [1eb20bca95](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1eb20bca95):